Add performance report DTOs and refactor report logic
Introduces new DTOs for evaluation items, performance results, and student performance. Refactors report generation logic in controllers and services to use user context and new DTOs, improving separation of concerns and maintainability. Adds StaffHelper for staff/admin checks, updates attendance and grade DTOs, and streamlines PDF report generation.

+++ ctp-docente-portal.Server/DTOs/Students/StudentReportDto.cs
@@ -8,7 +8,7 @@ namespace ctp_docente_portal.Server.DTOs.Students
         public string Name { get; set; }
         public string Identification { get; set; }
         public SimpleDto Group { get; set; }
-        public decimal Average { get; set; }
+        public int Average { get; set; }
         public decimal Attendance { get; set; }
         public SimpleDto Status { get; set; }
     }
